Move async_get_calendars to a seprate file
